smodu-backend/requirements/base.txt
Django==5.2.*
djangorestframework==3.16.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.4.*
celery==5.5.*
redis==5.2.*
psycopg2-binary==2.9.*
Pillow==10.*
django-filter==24.*
drf-spectacular==0.27.*
django-celery-beat>=2.8.0
gunicorn==23.*